+++ src/main/java/net/minecraft/client/renderer/texture/TextureManager.java
@@ -117,7 +117,7 @@ public class TextureManager implements ITickable, IResourceManagerReloadListener
         if (integer == null) {
             integer = Integer.valueOf(1);
         } else {
-            integer = Integer.valueOf(integer.intValue() + 1);
+            integer = Integer.valueOf(integer + 1);
         }
 
         this.mapTextureCounters.put(name, integer);
