+++ src/main/java/net/minecraft/client/renderer/texture/TextureMap.java
@@ -210,7 +210,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                             Iterator iterator1 = list1.iterator();
 
                             while (iterator1.hasNext()) {
-                                int j4 = ((Integer) iterator1.next()).intValue();
+                                int j4 = (Integer) iterator1.next();
 
                                 if (j4 > 0 && j4 < abufferedimage.length - 1 && abufferedimage[j4] == null) {
                                     ResourceLocation resourcelocation = this.completeResourceLocation(resourcelocation1, j4);
@@ -653,7 +653,7 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
                                 if (!map.containsKey(Integer.valueOf(j))) {
                                     map.put(Integer.valueOf(j), Integer.valueOf(1));
                                 } else {
-                                    int k = ((Integer) map.get(Integer.valueOf(j))).intValue();
+                                    int k = (Integer) map.get(Integer.valueOf(j));
                                     map.put(Integer.valueOf(j), Integer.valueOf(k + 1));
                                 }
                             }
@@ -670,8 +670,8 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
         int l1;
 
         for (Iterator iterator = set1.iterator(); iterator.hasNext(); l += l1) {
-            int j1 = ((Integer) iterator.next()).intValue();
-            l1 = ((Integer) map.get(Integer.valueOf(j1))).intValue();
+            int j1 = (Integer) iterator.next();
+            l1 = (Integer) map.get(Integer.valueOf(j1));
         }
 
         int i1 = 16;
@@ -680,8 +680,8 @@ public class TextureMap extends AbstractTexture implements ITickableTextureObjec
         Iterator iterator1 = set1.iterator();
 
         while (iterator1.hasNext()) {
-            int i2 = ((Integer) iterator1.next()).intValue();
-            int j2 = ((Integer) map.get(Integer.valueOf(i2))).intValue();
+            int i2 = (Integer) iterator1.next();
+            int j2 = (Integer) map.get(Integer.valueOf(i2));
             k1 += j2;
 
             if (i2 > i1) {
