+++ src/main/java/net/minecraft/world/storage/MapStorage.java
@@ -135,13 +135,13 @@ public class MapStorage {
         if (oshort == null) {
             oshort = Short.valueOf((short) 0);
         } else {
-            oshort = Short.valueOf((short) (oshort.shortValue() + 1));
+            oshort = Short.valueOf((short) (oshort + 1));
         }
 
         this.idCounts.put(key, oshort);
 
         if (this.saveHandler == null) {
-            return oshort.shortValue();
+            return oshort;
         } else {
             try {
                 File file1 = this.saveHandler.getMapFileFromName("idcounts");
@@ -150,7 +150,7 @@ public class MapStorage {
                     NBTTagCompound nbttagcompound = new NBTTagCompound();
 
                     for (String s : this.idCounts.keySet()) {
-                        short short1 = this.idCounts.get(s).shortValue();
+                        short short1 = this.idCounts.get(s);
                         nbttagcompound.setShort(s, short1);
                     }
 
@@ -162,7 +162,7 @@ public class MapStorage {
                 exception.printStackTrace();
             }
 
-            return oshort.shortValue();
+            return oshort;
         }
     }
 }
